requirements.txt
streamlit>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.8.0
Pillow>=10.0.0
pyttsx3>=2.90
protobuf>=5.29.0
